+++ Cosette/Engine/Ai/Search/SearchStatistics.cs
@@ -57,6 +57,7 @@ namespace Cosette.Engine.Ai.Search
         public int IIDHits { get; set; }
         public int LoudMovesGenerated { get; set; }
         public int QuietMovesGenerated { get; set; }
+        public int Extensions { get; set; }
 
         public Move[] PrincipalVariation { get; set; }
         public int PrincipalVariationMovesCount { get; set; }